STOP

+++ HotelReservationSystem-ejb/src/java/ejb/session/stateless/ReservationEntitySessionBean.java
@@ -6,6 +6,7 @@
 package ejb.session.stateless;
 
 import entity.Occupant;
+import entity.Partner;
 import entity.Report;
 import entity.Reservation;
 import entity.Room;
@@ -19,7 +20,6 @@ import javax.persistence.EntityManager;
 import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
 import util.enumeration.RoomStatusEnum;
-import util.exception.ReportNotFoundException;
 import util.exception.ReservationNotFoundException;
 import util.exception.RoomNotFoundException;
 
@@ -64,8 +64,15 @@ public class ReservationEntitySessionBean implements ReservationEntitySessionBea
         Reservation reservation = em.find(Reservation.class, reservationId);
 
         if (reservation != null) {
+            reservation.getOccupant();
+            reservation.getPartner();
             reservation.getRoomType();
-
+            reservation.getRooms().size();
+            List<Room> rooms = reservation.getRooms();
+            for (Room r : rooms) {
+                r.getRoomType();
+            }
+            reservation.getReports().size();
             return reservation;
         } else {
             throw new ReservationNotFoundException("Reservation with ID: " + reservationId + " does not exist!");
@@ -91,7 +98,7 @@ public class ReservationEntitySessionBean implements ReservationEntitySessionBea
 
     @Override
     public List<Reservation> getAllReservations() {
-        
+
         List<Reservation> reservations = em.createQuery("SELECT r from Reservation r")
                 .getResultList();
 
@@ -102,7 +109,7 @@ public class ReservationEntitySessionBean implements ReservationEntitySessionBea
             r.getRoomType();
             r.getRooms().size();
         }
-       
+
         return reservations;
     }
 
@@ -146,12 +153,14 @@ public class ReservationEntitySessionBean implements ReservationEntitySessionBea
         // manage the reservation
         reservation = em.merge(reservation);
         int requiredQuantity = reservation.getQuantity();
-        List<Room> allocations = reservation.getRooms();
+        List<Room> allocations = new ArrayList<>();
+        allocations.addAll(reservation.getRooms());
 
         // check if already allocated previously for this day's reservation (same day check-in)
         if (allocations.size() > 0) {
             return;
         }
+        System.out.println("getting room.");
 
         // get available rooms of reservation's room type
         List<Room> requiredRooms = em.createQuery("SELECT r from Room r WHERE r.roomType = ?1 AND r.enabled = ?2 AND r.roomStatus = ?3")
@@ -159,6 +168,8 @@ public class ReservationEntitySessionBean implements ReservationEntitySessionBea
                 .setParameter(2, true)
                 .setParameter(3, RoomStatusEnum.AVAILABLE)
                 .getResultList();
+
+        System.out.println(requiredQuantity);
 
         // get available rooms of reservation's higher room type
         RoomType higherRoomType = reservation.getRoomType().getNextHigherRoomType();
@@ -170,6 +181,8 @@ public class ReservationEntitySessionBean implements ReservationEntitySessionBea
                     .setParameter(3, RoomStatusEnum.AVAILABLE)
                     .getResultList();
         }
+
+        System.out.println(requiredQuantity);
 
         // if enough required rooms available, allocate sequentially.
         if (requiredRooms.size() >= requiredQuantity) {
@@ -192,6 +205,7 @@ public class ReservationEntitySessionBean implements ReservationEntitySessionBea
                 higherRoom.setReservation(reservation);
                 higherRoom.setRoomStatus(RoomStatusEnum.NOT_AVAILABLE);
                 Report report = new Report(1);
+                em.persist(report);
                 reservation.getReports().add(report);
                 requiredQuantity--;
             }
@@ -201,9 +215,13 @@ public class ReservationEntitySessionBean implements ReservationEntitySessionBea
         if (requiredQuantity > 0) {
             for (int i = 0; i < requiredQuantity; i++) {
                 Report report = new Report(2);
+                em.persist(report);
                 reservation.getReports().add(report);
             }
         }
+
+        // set allocations to reservation
+        reservation.setRooms(allocations);
 
     }
 
@@ -213,15 +231,36 @@ public class ReservationEntitySessionBean implements ReservationEntitySessionBea
             Room room = (Room) em.createQuery("SELECT r FROM Room r WHERE r.roomNumber = ?1")
                     .setParameter(1, roomNumber)
                     .getSingleResult();
-            room.setRoomStatus(RoomStatusEnum.NOT_AVAILABLE);
-            room.setReservation(null);
+
+            room.setRoomStatus(RoomStatusEnum.AVAILABLE);
             Reservation reservation = room.getReservation();
-            reservation.setPartner(null);
-            reservation.setOccupant(null);
-            reservation.setReports(null);
-            em.remove(reservation);
+            reservation.getRooms().remove(room);
+            room.setReservation(null);
+
+            // if after diassociating reservation and string for room checkout,
+            // no more rooms left to checkout in reservation,
+            // remove reservation.
+            if (reservation.getRooms().isEmpty()) {
+                // remove reservation record
+                Partner p = reservation.getPartner();
+                if (p != null) {
+                    reservation.setPartner(null);
+                    p.getReservations().remove(reservation);
+
+                }
+                Occupant o = reservation.getOccupant();
+                if (o != null) {
+                    reservation.setOccupant(null);
+                    o.getReservations().remove(reservation);
+                }
+
+                reservation.setRooms(null);
+                reservation.setReports(null);
+                em.remove(reservation);
+            }
+
         } catch (NoResultException ex) {
-            throw new RoomNotFoundException("This room does not exist!");
+            throw new RoomNotFoundException("This room does not exist!\n");
         }
     }
